requirements.txt
pip==21.3.1
setuptools==59.2.0
pytest-testinfra==6.4.0
cookiecutter==1.7.3
